--- a/src/lib.rs
+++ b/src/lib.rs
@@ -37,7 +37,7 @@
 //! #    run(filename).unwrap();
 //! # }
 //!
-//! fn parse_rl<'a>(data: &'a [u8]) -> Result<Replay<'a>, ParseError> {
+//! fn parse_rl<'a>(data: &'a [u8]) -> Result<Replay, ParseError> {
 //!     boxcars::ParserBuilder::new(&data)
 //!         .on_error_check_crc()
 //!         .parse()