--- a/src/network/attributes.rs
+++ b/src/network/attributes.rs
@@ -1009,11 +1009,11 @@ fn decode_text(bits: &mut BitGet<'_>) -> Result<String, AttributeError> {
             .checked_mul(-2)
             .ok_or_else(|| AttributeError::TooBigString(size))?;
         bits.read_bytes(len)
-            .and_then(|data| decode_utf16(&data[..]).map(Cow::into_owned).ok())
+            .and_then(|data| decode_utf16(&data[..]).ok())
             .ok_or_else(|| AttributeError::TooBigString(len))
     } else {
         bits.read_bytes(size)
-            .and_then(|data| decode_windows1252(&data[..]).map(Cow::into_owned).ok())
+            .and_then(|data| decode_windows1252(&data[..]).ok())
             .ok_or_else(|| AttributeError::TooBigString(size))
     }
 }