Header takes longer to parse but body is faster overall.  I still need to fix names example.