@nickbabcock 2019-09-08
Fascinating, I can corroborate your findings that these changes improve performance when parsing the whole replay (header + network). A ~4-5% performance improvement (from 19ms to 18.5ms). Inexplicable, but welcomed.

I did experience a 66% increase in the time it takes to just parse the header (from 56us to 95us). While this appears bad at first glance, this PR still nets a 500us performance win ((19ms - 18.5ms) > (95us - 56us)).

imo the fraction of a millisecond cost in header-only parsing is worth the ergonomic improvements.

Let me know when you've updated the names example and I can get this merged and released.